requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-pinecone
pinecone-client